Moved yelp server files to own folder
